chore: Rename templating -> template
Originally implemented in and imported from https://github.com/stackabletech/config-utils
